add UserModel, change start activity

+++ app/src/main/java/com/example/mvptry/common/User.kt
@@ -1,3 +1,3 @@
 package com.example.mvptry.common
 
-data class User(var id: Int, var name: String, var email: String)
+data class User(var id: Long, var name: String, var email: String)
